#310 Rename: ディレクトリ名を変更
